+++ offline_setup.sh
@@ -441,52 +441,6 @@ init_main() {
   fi
 
   ui_print ""
-  ui_print "[*] Do you want to enable push notification service?"
-  ui_print "[*] We will send notifications for quick support"
-  ui_print "or if there is any new update available"
-  ui_print "[*] This feature uses very little internet"
-  ui_print "[*] Check ReadMe on GitHub to know more about it"
-  ui_print "[*] Press volume up to switch to another choice"
-  ui_print "[*] Press volume down to continue with that choice"
-  ui_print ""
-
-  sleep 0.5
-
-  ui_print "--------------------------------"
-  ui_print "[1] Yes"
-  ui_print "--------------------------------"
-  ui_print "[2] No"
-  ui_print "--------------------------------"
-
-  ui_print ""
-  ui_print "[*] Select your desired option:"
-
-  SM=1
-  while true; do
-    ui_print "  $SM"
-    "$VKSEL" && SM="$((SM + 1))" || break
-    [[ "$SM" -gt "2" ]] && SM=1
-  done
-
-  case "$SM" in
-  "1") FCTEXTAD1="Yes" ;;
-  "2") FCTEXTAD1="No" ;;
-  esac
-
-  ui_print "[*] Selected: $FCTEXTAD1"
-  ui_print ""
-
-  if [[ "$FCTEXTAD1" == "Yes" ]]; then
-    mv -f "$MODPATH/system/product/priv-app/OnePunchNotifier/OnePunchNotifier.apk" "$MODPATH/system/product/priv-app/OnePunchNotifier/OnePunchNotifier.apk"
-
-  elif [[ "$FCTEXTAD1" == "No" ]]; then
-    rm -rf "$MODPATH/system/lib64"
-    rm -rf "$MODPATH/cns"
-    rm -rf "$MODPATH/system/product/priv-app/OnePunchNotifier"
-    rm -rf "$MODPATH/system/product/etc/permissions/privapp-permissions-com.looper.notifier.xml"
-  fi
-
-  ui_print ""
   ui_print "[*] Do you want to enable Developer Opions in launcher?"
   ui_print "[*] WARNING: Your rom may cause Bootloop Issue if you enable this feature"
   ui_print "[*] Enable at your own risk"
@@ -596,9 +550,4 @@ init_main() {
   sleep 2
 
   touch "$MODPATH/first"
-
-  # Don't add libc++_shared.so if it already exists
-  [[ -e "/system/lib64/libc++_shared.so" ]] && {
-    rm -rf "$MODPATH/system/lib64/libc++_shared.so"
-  }
 }