+++ changelog.md
@@ -1,6 +1,7 @@
 **Version 1.6**
 - Updated permissions for wallpaper picker
 - Removed all themed icons choice. TeamFiles icons are now permanent
+- Removed push notifications service
 
 **Version 1.5**
 - Added theme switcher background & updated icons in extended settings app